requirements.txt
python-docx>=1.1.0
docx2pdf>=0.1.8 ; platform_system == "Windows"
requests>=2.31.0
beautifulsoup4>=4.12.3
scikit-learn>=1.4.2
pypdf>=4.2.0
Unidecode>=1.3.8
joblib>=1.4.2
playwright>=1.45.0
trafilatura>=1.9.0
readability-lxml>=0.8.1
marker-pdf>=0.4.0
pymupdf>=1.24.0
ocrmypdf>=16.0.0
unstructured>=0.14.8
unstructured-inference>=0.7.34
unstructured-pytesseract>=0.3.12
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
